requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy>=1.23.0
requests>=2.31.0
torch>=2.0.0
